+++ spec/integrations/pro/routing/virtual_partitions_with_manual_offset_management.rb
@@ -1,0 +1,28 @@
+# frozen_string_literal: true
+
+# Karafka should not allow for using VP with MOM
+
+setup_karafka do |config|
+  config.license.token = pro_license_token
+  config.concurrency = 10
+end
+
+guarded = []
+
+begin
+  draw_routes do
+    consumer_group DT.consumer_group do
+      topic DT.topic do
+        consumer Class.new(Karafka::Pro::BaseConsumer)
+        virtual_partitions(
+          partitioner: ->(msg) { msg.raw_payload }
+        )
+        manual_offset_management true
+      end
+    end
+  end
+rescue Karafka::Errors::InvalidConfigurationError
+  guarded << 1
+end
+
+assert_equal [1], guarded