+++ spec/lib/karafka/base_consumer_spec.rb
@@ -20,10 +20,10 @@ RSpec.describe_current do
       Karafka::Messages::Messages,
       first: first_message,
       last: last_message,
-      metadata: instance_double(
-        Karafka::Messages::BatchMetadata,
+      metadata: Karafka::Messages::BatchMetadata.new(
         topic: topic.name,
-        partition: 0
+        partition: 0,
+        processed_at: Time.now
       )
     )
   end
@@ -61,12 +61,14 @@ RSpec.describe_current do
   describe '#on_consume and #on_after_consume' do
     let(:consume_with_after) do
       lambda do
+        consumer.on_before_consume
         consumer.on_consume
         consumer.on_after_consume
       end
     end
 
     before do
+      coordinator.seek_offset = messages.first.offset
       consumer.coordinator = coordinator
       consumer.client = client
       consumer.messages = messages
@@ -106,7 +108,7 @@ RSpec.describe_current do
 
       it { expect { consume_with_after.call }.not_to raise_error }
 
-      it 'expect to pause based on the message offset' do
+      it 'expect to pause based on the first message ever received' do
         consume_with_after.call
         expect(client).to have_received(:pause).with(topic.name, first_message.partition, offset)
       end
@@ -289,7 +291,7 @@ RSpec.describe_current do
       end
 
       it 'epxect to increase seek_offset' do
-        expect(consumer.instance_variable_get(:@seek_offset)).to eq(offset + 1)
+        expect(consumer.coordinator.seek_offset).to eq(offset + 1)
       end
     end
 
@@ -305,7 +307,7 @@ RSpec.describe_current do
       end
 
       it 'epxect to not increase seek_offset' do
-        expect(consumer.instance_variable_get(:@seek_offset)).to eq(nil)
+        expect(consumer.coordinator.seek_offset).to eq(nil)
       end
     end
   end
@@ -325,7 +327,7 @@ RSpec.describe_current do
       end
 
       it 'epxect to increase seek_offset' do
-        expect(consumer.instance_variable_get(:@seek_offset)).to eq(offset + 1)
+        expect(consumer.coordinator.seek_offset).to eq(offset + 1)
       end
     end
 
@@ -341,7 +343,7 @@ RSpec.describe_current do
       end
 
       it 'epxect to not increase seek_offset' do
-        expect(consumer.instance_variable_get(:@seek_offset)).to eq(nil)
+        expect(consumer.coordinator.seek_offset).to eq(nil)
       end
     end
   end